% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\docType{data}
\name{data1}
\alias{data1}
\title{data1: the namespace variable.}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
data1
}
\description{
the namespace variable that stores data provided by the user.
}
\keyword{datasets}
